backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy[asyncio]==2.0.25
aiosqlite==0.19.0
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
pyotp==2.9.0
qrcode[pil]==7.4.2
python-multipart==0.0.9
httpx==0.26.0
email-validator==2.1.0
pydantic-settings==2.1.0
slowapi==0.1.9
